refactor(subtitle): move some function to helpers  folder
